backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.3
celery==5.3.6
transformers==4.38.2
sentencepiece==0.2.0
huggingface-hub==0.21.4
groq==0.4.2
anthropic==0.20.0
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1
python-multipart==0.0.9
reportlab==4.1.0
